+++ src/main/java/com/bootcamp/be_java_hisp_w25_g02/dto/response/UserFollowingDTO.java
@@ -3,7 +3,7 @@ package com.bootcamp.be_java_hisp_w25_g02.dto.response;
 import java.util.List;
 
 public record UserFollowingDTO(
-        Long user_id,
+        Integer user_id,
         String user_name,
         List<UserDTO> followed
 ) {
