+++ src/main/java/com/bootcamp/be_java_hisp_w25_g02/service/IUserService.java
@@ -5,5 +5,5 @@ import com.bootcamp.be_java_hisp_w25_g02.dto.response.UserFollowingDTO;
 import java.util.List;
 
 public interface IUserService {
-    UserFollowingDTO getFollowedSellers(Long userId);
+    UserFollowingDTO getFollowedSellers(Integer userId);
 }
